modified: rename extensions to ApiDependencyInjections
