+++ MBS.API/ApiDependencyInjections/ServicesDependencyInjection.cs
@@ -1,7 +1,7 @@
 ﻿using MBS.Application.Services.Implements;
 using MBS.Application.Services.Interfaces;
 
-namespace MBS.API.Extensions;
+namespace MBS.API.ApiDependencyInjections;
 
 public static class ServicesDependencyInjection
 {
